Move all docker and external db code to `piker.service`
