+++ piker/cli/__init__.py
@@ -206,8 +206,8 @@ def services(config, tl, ports):
 
 
 def _load_clis() -> None:
-    from ..data import marketstore  # noqa
-    from ..data import elastic
+    from ..service import marketstore  # noqa
+    from ..service import elastic
     from ..data import cli  # noqa
     from ..brokers import cli  # noqa
     from ..ui import cli  # noqa
